Update BoringSSL to abfd5ebc87ddca0fab9fca067c9d7edbc355eae8 (#424)
* Update BoringSSL vendoring scripts for new perlasm

* More patches

* Update BoringSSL to abfd5ebc87ddca0fab9fca067c9d7edbc355eae8

* Fixup newly opaquified types

+++ Sources/CNIOBoringSSL/crypto/asn1/posix_time.c
@@ -15,6 +15,8 @@
 // Time conversion to/from POSIX time_t and struct tm, with no support
 // for time zones other than UTC
 
+#include <CNIOBoringSSL_time.h>
+
 #include <assert.h>
 #include <inttypes.h>
 #include <limits.h>
@@ -191,9 +193,10 @@ int OPENSSL_gmtime_adj(struct tm *tm, int off_day, long offset_sec) {
                            tm->tm_hour, tm->tm_min, tm->tm_sec, &posix_time)) {
     return 0;
   }
-  if (!utc_from_posix_time(posix_time + off_day * SECS_PER_DAY + offset_sec,
-                           &tm->tm_year, &tm->tm_mon, &tm->tm_mday,
-                           &tm->tm_hour, &tm->tm_min, &tm->tm_sec)) {
+  if (!utc_from_posix_time(
+          posix_time + (int64_t)off_day * SECS_PER_DAY + offset_sec,
+          &tm->tm_year, &tm->tm_mon, &tm->tm_mday, &tm->tm_hour, &tm->tm_min,
+          &tm->tm_sec)) {
     return 0;
   }
   tm->tm_year -= 1900;
